- Fixed an error caused by using byte as the base type in Byte.java - Expanded on simpletest.asm - Added PC incrementation. - Added some nice output.

+++ Byte.java
@@ -8,7 +8,7 @@
 public class Byte
 {
     // This class stores a signed byte.
-    private byte value;
+    private int value;
     
     /**
      * Creates a default Byte ( value = 0 )
